Move lock registration to separate module
Remove global lock registration as it's unused.
