+++ libs/core/concurrency/CMakeLists.txt
@@ -51,6 +51,7 @@ add_hpx_module(
     hpx_errors
     hpx_hashing
     hpx_itt_notify
+    hpx_lock_registration
     hpx_thread_support
   CMAKE_SUBDIRS examples tests
 )
