+++ libs/core/init_runtime_local/CMakeLists.txt
@@ -29,6 +29,7 @@ add_hpx_module(
     hpx_errors
     hpx_filesystem
     hpx_format
+    hpx_lock_registration
     hpx_logging
     hpx_schedulers
     hpx_testing
