+++ libs/core/lcos_local/include/hpx/lcos_local/channel.hpp
@@ -9,11 +9,11 @@
 #include <hpx/config.hpp>
 #include <hpx/assert.hpp>
 #include <hpx/async_base/launch_policy.hpp>
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/futures/future.hpp>
 #include <hpx/futures/packaged_task.hpp>
 #include <hpx/iterator_support/iterator_facade.hpp>
 #include <hpx/lcos_local/receive_buffer.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/memory.hpp>
 #include <hpx/synchronization/no_mutex.hpp>
