+++ libs/core/lock_registration/README.rst
@@ -1,0 +1,16 @@
+
+..
+    Copyright (c) 2020-2021 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+=================
+lock_registration
+=================
+
+This module is part of HPX.
+
+Documentation can be found `here
+<https://hpx-docs.stellar-group.org/latest/html/modules/lock_registration/docs/index.html>`__.
