+++ libs/core/lock_registration/docs/index.rst
@@ -1,0 +1,19 @@
+..
+    Copyright (c) 2020-2021 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _modules_lock_registration:
+
+=================
+lock_registration
+=================
+
+This module contains fucntionality for registering locks to detect when they are
+locked and unlocked on different threads.
+
+See the :ref:`API reference <modules_lock_registration_api>` of this module for more
+details.
+
