+++ libs/core/lock_registration/tests/CMakeLists.txt
@@ -1,0 +1,42 @@
+# Copyright (c) 2020-2021 The STE||AR-Group
+#
+# SPDX-License-Identifier: BSL-1.0
+# Distributed under the Boost Software License, Version 1.0. (See accompanying
+# file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+include(HPX_Message)
+
+if(HPX_WITH_TESTS)
+  if(HPX_WITH_TESTS_UNIT)
+    add_hpx_pseudo_target(tests.unit.modules.lock_registration)
+    add_hpx_pseudo_dependencies(
+      tests.unit.modules tests.unit.modules.lock_registration
+    )
+    add_subdirectory(unit)
+  endif()
+
+  if(HPX_WITH_TESTS_REGRESSIONS)
+    add_hpx_pseudo_target(tests.regressions.modules.lock_registration)
+    add_hpx_pseudo_dependencies(
+      tests.regressions.modules tests.regressions.modules.lock_registration
+    )
+    add_subdirectory(regressions)
+  endif()
+
+  if(HPX_WITH_TESTS_BENCHMARKS)
+    add_hpx_pseudo_target(tests.performance.modules.lock_registration)
+    add_hpx_pseudo_dependencies(
+      tests.performance.modules tests.performance.modules.lock_registration
+    )
+    add_subdirectory(performance)
+  endif()
+
+  if(HPX_WITH_TESTS_HEADERS)
+    add_hpx_header_tests(
+      modules.lock_registration
+      HEADERS ${lock_registration_headers}
+      HEADER_ROOT ${PROJECT_SOURCE_DIR}/include
+      DEPENDENCIES hpx_lock_registration
+    )
+  endif()
+endif()
