+++ libs/core/modules.rst
@@ -51,6 +51,7 @@ Core modules
    /libs/core/iterator_support/docs/index.rst
    /libs/core/itt_notify/docs/index.rst
    /libs/core/lcos_local/docs/index.rst
+   /libs/core/lock_registration/docs/index.rst
    /libs/core/logging/docs/index.rst
    /libs/core/memory/docs/index.rst
    /libs/core/mpi_base/docs/index.rst
