+++ libs/core/runtime_configuration/CMakeLists.txt
@@ -13,7 +13,6 @@ set(runtime_configuration_headers
     hpx/runtime_configuration/component_registry_base.hpp
     hpx/runtime_configuration/init_ini_data.hpp
     hpx/runtime_configuration/plugin_registry_base.hpp
-    hpx/runtime_configuration/register_locks_globally.hpp
     hpx/runtime_configuration/runtime_configuration.hpp
     hpx/runtime_configuration/runtime_configuration_fwd.hpp
     hpx/runtime_configuration/runtime_mode.hpp
@@ -25,7 +24,6 @@ set(
   runtime_configuration_compat_headers
   hpx/util/ini.hpp => hpx/modules/runtime_configuration.hpp
   hpx/util/init_ini_data.hpp => hpx/modules/runtime_configuration.hpp
-  hpx/util/register_locks_globally.hpp => hpx/modules/runtime_configuration.hpp
   hpx/util/runtime_configuration.hpp => hpx/modules/runtime_configuration.hpp
   hpx/plugins/plugin_registry_base.hpp => hpx/modules/runtime_configuration.hpp
   hpx/runtime/components/component_commandline_base.hpp => hpx/modules/runtime_configuration.hpp
@@ -36,9 +34,8 @@ set(
 )
 # cmake-format: on
 
-set(runtime_configuration_sources
-    init_ini_data.cpp register_locks_globally.cpp runtime_configuration.cpp
-    runtime_mode.cpp static_factory_data.cpp
+set(runtime_configuration_sources init_ini_data.cpp runtime_configuration.cpp
+                                  runtime_mode.cpp static_factory_data.cpp
 )
 
 include(HPX_AddModule)
