+++ libs/core/runtime_configuration/src/runtime_configuration.cpp
@@ -128,13 +128,6 @@ namespace hpx { namespace util {
 #endif
             "throw_on_held_lock = ${HPX_THROW_ON_HELD_LOCK:1}",
 #endif
-#ifdef HPX_HAVE_VERIFY_LOCKS_GLOBALLY
-#if defined(HPX_DEBUG)
-            "global_lock_detection = ${HPX_GLOBAL_LOCK_DETECTION:1}",
-#else
-            "global_lock_detection = ${HPX_GLOBAL_LOCK_DETECTION:0}",
-#endif
-#endif
 #ifdef HPX_HAVE_THREAD_MINIMAL_DEADLOCK_DETECTION
 #ifdef HPX_DEBUG
             "minimal_deadlock_detection = ${HPX_MINIMAL_DEADLOCK_DETECTION:1}",
@@ -940,19 +933,6 @@ namespace hpx { namespace util {
         if (util::section const* sec = get_section("hpx"); nullptr != sec)
         {
             return hpx::util::get_entry_as<int>(*sec, "lock_detection", 0) != 0;
-        }
-#endif
-        return false;
-    }
-
-    // Enable global lock tracking
-    bool runtime_configuration::enable_global_lock_detection() const
-    {
-#ifdef HPX_HAVE_VERIFY_LOCKS_GLOBALLY
-        if (util::section const* sec = get_section("hpx"); nullptr != sec)
-        {
-            return hpx::util::get_entry_as<int>(
-                       *sec, "global_lock_detection", 0) != 0;
         }
 #endif
         return false;
