+++ libs/core/synchronization/CMakeLists.txt
@@ -77,6 +77,7 @@ add_hpx_module(
     hpx_hashing
     hpx_itt_notify
     hpx_memory
+    hpx_lock_registration
     hpx_logging
     hpx_memory
     hpx_threading_base
