+++ libs/core/synchronization/include/hpx/synchronization/spinlock.hpp
@@ -15,8 +15,8 @@
 
 #include <hpx/config.hpp>
 
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/execution_base/this_thread.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/modules/itt_notify.hpp>
 
 #include <atomic>
