+++ libs/core/synchronization/src/mutex.cpp
@@ -9,7 +9,7 @@
 
 #include <hpx/assert.hpp>
 #include <hpx/coroutines/thread_enums.hpp>
-#include <hpx/execution_base/register_locks.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/itt_notify.hpp>
 #include <hpx/synchronization/condition_variable.hpp>
