+++ libs/core/threading_base/include/hpx/threading_base/thread_helpers.hpp
@@ -11,8 +11,8 @@
 
 #include <hpx/config.hpp>
 #include <hpx/coroutines/thread_enums.hpp>
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/functional/unique_function.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/threading_base/register_thread.hpp>
 #include <hpx/threading_base/scheduler_mode.hpp>
