+++ libs/core/threading_base/src/thread_helpers.cpp
@@ -11,7 +11,7 @@
 #include <hpx/coroutines/thread_enums.hpp>
 #include <hpx/modules/errors.hpp>
 #ifdef HPX_HAVE_VERIFY_LOCKS
-#include <hpx/execution_base/register_locks.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #endif
 #include <hpx/execution_base/this_thread.hpp>
 #include <hpx/threading_base/detail/reset_lco_description.hpp>
