+++ libs/full/CMakeLists.txt
@@ -41,7 +41,7 @@ set(_hpx_full_modules
 # cmake-format: on
 
 hpx_info("")
-hpx_info("  Configuring libhpx modules")
+hpx_info("  Configuring libhpx modules:")
 
 foreach(module ${_hpx_full_modules})
   add_subdirectory(${module})
