+++ libs/full/agas_base/src/server/primary_namespace_server.cpp
@@ -12,8 +12,8 @@
 #include <hpx/async_distributed/applier/apply.hpp>
 #include <hpx/async_distributed/continuation.hpp>
 #include <hpx/components_base/agas_interface.hpp>
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/format.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/modules/async_distributed.hpp>
 #include <hpx/modules/errors.hpp>
 #include <hpx/modules/logging.hpp>
