+++ libs/full/collectives/include/hpx/collectives/detail/channel_communicator.hpp
@@ -15,9 +15,9 @@
 #include <hpx/components/client.hpp>
 #include <hpx/components_base/server/component_base.hpp>
 #include <hpx/datastructures/any.hpp>
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/futures/future.hpp>
 #include <hpx/lcos_local/channel.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/synchronization/spinlock.hpp>
 #include <hpx/type_support/unused.hpp>
