+++ libs/full/components_base/include/hpx/components_base/server/locking_hook.hpp
@@ -10,8 +10,8 @@
 #include <hpx/components_base/get_lva.hpp>
 #include <hpx/components_base/traits/action_decorate_function.hpp>
 #include <hpx/coroutines/coroutine.hpp>
-#include <hpx/execution_base/register_locks.hpp>
 #include <hpx/functional/bind_front.hpp>
+#include <hpx/lock_registration/detail/register_locks.hpp>
 #include <hpx/synchronization/spinlock.hpp>
 #include <hpx/thread_support/unlock_guard.hpp>
 #include <hpx/threading_base/thread_data.hpp>
